FM-41: Move stores to own module.
