+++ fendermint/vm/interpreter/src/fvm/store/memory.rs
